defi-analyser/requirements.txt.txt
# Core dependencies
web3>=6.0.0
eth-typing>=3.0.0
eth-utils>=2.1.0
eth-abi>=3.0.0
asyncio>=3.4.3
aiohttp>=3.8.1

# Security Analysis Tools
slither-analyzer>=0.9.0
mythril>=0.23.15
echidna-test>=2.0.0
manticore>=0.3.7
crytic-compile>=0.2.3

# Development Frameworks
foundry-rs>=0.2.0  # For Foundry
hardhat>=2.12.0    # For Hardhat
eth-ape>=0.5.0     # For Ape
eth-brownie>=1.19.0  # For Brownie

# Monitoring and Deployment
tenderly>=0.9.0

# Testing and Analysis
pytest>=7.0.0
pytest-asyncio>=0.18.0
hypothesis>=6.0.0

# AI/ML Components
torch>=1.9.0
transformers>=4.30.0
numpy>=1.21.0

# Monitoring and Logging
prometheus-client>=0.14.0
python-json-logger>=2.0.0

# Development Tools
black>=22.0.0
mypy>=0.950
pylint>=2.14.0

# Database
sqlalchemy>=1.4.0
aiosqlite>=0.17.0

# Utilities
python-dotenv>=0.19.0
dataclasses-json>=0.5.0
typing-extensions>=4.0.0